fix(sonarqube): SonarQube fix for Dev to Main Release 2.98
