+++ src/scripts/scriptUtil.js
@@ -1,0 +1,13 @@
+const readline = require('node:readline');
+
+function prompt(question) {
+  const rl = readline.createInterface({ input: process.stdin, output: process.stdout });
+  return new Promise((resolve) => {
+    rl.question(question, (answer) => {
+      rl.close();
+      resolve(answer);
+    });
+  });
+}
+
+module.exports = { prompt };
